Move Inhertiance into own class.
