+++ org.eclipse.winery.frontends/app/topologymodeler/src/app/canvas/canvas.component.ts
@@ -58,6 +58,7 @@ import { WineryRepositoryConfigurationService } from '../../../../tosca-manageme
 import { RequirementDefinitionModel } from '../models/requirementDefinitonModel';
 import { CapabilityDefinitionModel } from '../models/capabilityDefinitionModel';
 import { WineryRowData } from '../../../../tosca-management/src/app/wineryTableModule/wineryTable.component';
+import { InheritanceUtils } from '../models/InheritanceUtils';
 
 @Component({
     selector: 'winery-canvas',
@@ -396,7 +397,7 @@ export class CanvasComponent implements OnInit, OnDestroy, OnChanges, AfterViewI
                     try {
                         // request all valid requirement types for that node type for display as name select options in
                         // the modal
-                        const data = TopologyTemplateUtil.getRequirementDefinitionsOfNodeType(currentNodeData.type, this.entityTypes);
+                        const data = InheritanceUtils.getRequirementDefinitionsOfNodeType(currentNodeData.type, this.entityTypes);
                         this.requirements.reqDefinitionNames = [];
                         this.requirements.reqDefinitionName = '';
 
@@ -478,7 +479,7 @@ export class CanvasComponent implements OnInit, OnDestroy, OnChanges, AfterViewI
                     try {
                         // request all valid capability types for that node type for display as name select options in
                         // the modal
-                        const data = TopologyTemplateUtil.getCapabilityDefinitionsOfNodeType(currentNodeData.type, this.entityTypes);
+                        const data = InheritanceUtils.getCapabilityDefinitionsOfNodeType(currentNodeData.type, this.entityTypes);
                         this.capabilities.capDefinitionNames = [];
                         this.capabilities.capDefinitionName = '';
                         for (const capType of data) {
@@ -927,7 +928,7 @@ export class CanvasComponent implements OnInit, OnDestroy, OnChanges, AfterViewI
                                 && nt.requirements.requirement.some(req => req.id === reqId));
                         const requirementModel: RequirementModel = sourceNodeTemplate.requirements.requirement
                             .find(req => req.id === reqId);
-                        const requirementDefinition: RequirementDefinitionModel = TopologyTemplateUtil
+                        const requirementDefinition: RequirementDefinitionModel = InheritanceUtils
                             .getRequirementDefinitionsOfNodeType(sourceNodeTemplate.type, this.entityTypes)
                             .find(reqDef => reqDef.name === requirementModel.name);
                         requirementModel.capability = requirementDefinition.capability;
@@ -1869,7 +1870,7 @@ export class CanvasComponent implements OnInit, OnDestroy, OnChanges, AfterViewI
         if (this.configuration.isYaml()) {
             this.newNode.requirements = { requirement: [] };
             this.newNode.capabilities = { capability: [] };
-            const reqData = TopologyTemplateUtil.getRequirementDefinitionsOfNodeType(this.newNode.type, this.entityTypes);
+            const reqData = InheritanceUtils.getRequirementDefinitionsOfNodeType(this.newNode.type, this.entityTypes);
             if (reqData) {
                 reqData.forEach(reqDef => {
                     const reqModel = RequirementModel.fromRequirementDefinition(reqDef);
@@ -1877,7 +1878,7 @@ export class CanvasComponent implements OnInit, OnDestroy, OnChanges, AfterViewI
                     this.newNode.requirements.requirement.push(reqModel);
                 });
             }
-            const capData = TopologyTemplateUtil.getCapabilityDefinitionsOfNodeType(this.newNode.type, this.entityTypes);
+            const capData = InheritanceUtils.getCapabilityDefinitionsOfNodeType(this.newNode.type, this.entityTypes);
             if (capData) {
                 capData.forEach(capDef => {
                     const capModel = CapabilityModel.fromCapabilityDefinitionModel(capDef);
@@ -2186,7 +2187,6 @@ export class CanvasComponent implements OnInit, OnDestroy, OnChanges, AfterViewI
                     const capabilityId: string = info.targetId.substring(info.targetId.indexOf('.') + 1);
                     const currentTypeValid = this.entityTypes.relationshipTypes.some(relType => relType.qName === this.selectedRelationshipType.qName);
                     const currentSourceIdValid = this.allNodeTemplates.some(node => node.id === sourceNode);
-
                     if (sourceNode && currentTypeValid && currentSourceIdValid) {
                         const prefix = this.backendService.configuration.relationshipPrefix;
                         const relName = this.selectedRelationshipType.name;
@@ -2204,10 +2204,10 @@ export class CanvasComponent implements OnInit, OnDestroy, OnChanges, AfterViewI
                         const targetNodeTypeString: string = this.allNodeTemplates
                             .filter(nodeTemplate => nodeTemplate.id === info.targetId.substring(0, info.targetId.indexOf('.')))
                             .map(nodeTemplate => nodeTemplate.type)[0];
-                        const capDef: CapabilityDefinitionModel = TopologyTemplateUtil
+                        const capDef: CapabilityDefinitionModel = InheritanceUtils
                             .getCapabilityDefinitionsOfNodeType(targetNodeTypeString, this.entityTypes)
                             .filter(current => current.name === capModel.name)[0];
-                        const reqDef: RequirementDefinitionModel = TopologyTemplateUtil
+                        const reqDef: RequirementDefinitionModel = InheritanceUtils
                             .getRequirementDefinitionsOfNodeType(sourceNodeTypeString, this.entityTypes)
                             .filter(current => current.name === reqModel.name)[0];
 
@@ -2314,7 +2314,7 @@ export class CanvasComponent implements OnInit, OnDestroy, OnChanges, AfterViewI
      */
     matchType(requiredType: string, targetElementType: string, targetElementTypeSet: EntityType[]) {
         if (requiredType) {
-            const typeAncestry = TopologyTemplateUtil.getInheritanceAncestry(targetElementType, targetElementTypeSet);
+            const typeAncestry = InheritanceUtils.getInheritanceAncestry(targetElementType, targetElementTypeSet);
             return typeAncestry.some(type => type.qName === requiredType);
         } else {
             // if there is no required type, we assume all target types are valid!
@@ -2386,7 +2386,7 @@ export class CanvasComponent implements OnInit, OnDestroy, OnChanges, AfterViewI
 
     handleYamlPolicySelected($event: WineryRowData) {
         this.selectedYamlPolicy = this.entityTypes.yamlPolicies.find(policy => policy.name === (<TPolicy>$event.row).name);
-        this.selectedYamlPolicy.properties = TopologyTemplateUtil.getActiveKVPropertiesOfTemplateElement(this.selectedYamlPolicy.properties,
+        this.selectedYamlPolicy.properties = InheritanceUtils.getEffectiveKVPropertiesOfTemplateElement(this.selectedYamlPolicy.properties,
             this.selectedYamlPolicy.policyType, this.entityTypes.policyTypes);
     }
